Keep types the primary source and update imports
