+++ frontend/src/lib/sends/recipients.ts
@@ -1,4 +1,4 @@
-import type { SendRecipient } from './types';
+import type { SendRecipient } from '../../routes/console/types';
 
 export function getSortedRecipients(recipients: SendRecipient[]) {
